refactor: split native hotspot helpers
