+++ native/rust/crates/ripdpi-config/src/model/defaults/quic.rs
@@ -1,0 +1,7 @@
+use super::{QuicInitialMode, RuntimeQuicSettings};
+
+impl Default for RuntimeQuicSettings {
+    fn default() -> Self {
+        Self { initial_mode: QuicInitialMode::RouteAndCache, support_v1: true, support_v2: true }
+    }
+}
