+++ native/rust/crates/ripdpi-monitor-engine/src/engine/runners/strategy/support/candidate_ordering.rs
@@ -1,10 +1,19 @@
-use std::time::Duration;
-
-use ripdpi_monitor_adapter::failure::{ClassifiedFailure, FailureClass};
+use ripdpi_monitor_adapter::failure::FailureClass;
 
 use crate::candidates::{CandidateEligibility, StrategyCandidateSpec};
 use crate::classification::{interleave_candidate_families, reorder_tcp_candidates_for_failure};
 use crate::types::ProbeResult;
+
+mod rst_timeout;
+mod tls_ech;
+mod tls_version_split;
+
+pub(in crate::engine::runners::strategy) use rst_timeout::compute_rst_adaptive_timeout;
+#[cfg(test)]
+pub(in crate::engine::runners::strategy) use tls_ech::baseline_has_tls_ech_only;
+pub(in crate::engine::runners::strategy) use tls_ech::baseline_supports_ech_candidates;
+pub(in crate::engine::runners::strategy) use tls_version_split::baseline_has_tls_version_split;
+use tls_version_split::promote_tls_version_split_families;
 
 pub(in crate::engine::runners::strategy) const ECH_ELIGIBILITY_RATIONALE: &str =
     "Baseline did not expose an ECH-capable HTTPS target, so ECH extension splitting would be a no-op";
@@ -51,66 +60,6 @@ impl<'a> FamilyFailureTracker<'a> {
     pub(in crate::engine::runners::strategy) fn blocked_family(&self) -> Option<&'a str> {
         self.blocked
     }
-}
-
-fn probe_detail_value<'a>(result: &'a ProbeResult, key: &str) -> Option<&'a str> {
-    result.details.iter().find(|detail| detail.key == key).map(|detail| detail.value.as_str())
-}
-
-pub(in crate::engine::runners::strategy) fn compute_rst_adaptive_timeout(
-    baseline_failure: &ClassifiedFailure,
-) -> Option<Duration> {
-    if !matches!(baseline_failure.class, FailureClass::TcpReset) {
-        return None;
-    }
-    Some(Duration::from_millis(1500))
-}
-
-#[cfg(test)]
-pub(in crate::engine::runners::strategy) fn baseline_has_tls_ech_only(results: &[ProbeResult]) -> bool {
-    results.iter().any(|result| result.probe_type == "strategy_https" && result.outcome == "tls_ech_only")
-}
-
-pub(in crate::engine::runners::strategy) fn baseline_supports_ech_candidates(results: &[ProbeResult]) -> bool {
-    results.iter().any(|result| {
-        result.probe_type == "strategy_https"
-            && (result.outcome == "tls_ech_only"
-                || probe_detail_value(result, "tlsEchResolutionDetail") == Some("ech_config_available")
-                || probe_detail_value(result, "cdnProvider").is_some_and(|value| !value.trim().is_empty()))
-    })
-}
-
-/// Split-probing signal consumer: the HTTPS baseline saw exactly one TLS
-/// version reachable while the other was blocked (the `tls_version_split`
-/// outcome — see the per-version probe attempts in the HTTPS lane). That is
-/// evidence of version-sensitive, SNI-based blocking, for which the
-/// ClientHello/TLS-record splitting families are the natural first responders.
-pub(in crate::engine::runners::strategy) fn baseline_has_tls_version_split(results: &[ProbeResult]) -> bool {
-    results.iter().any(|result| result.probe_type == "strategy_https" && result.outcome == "tls_version_split")
-}
-
-/// SNI-splitting follow-up families, promoted first when the baseline reports a
-/// `tls_version_split`. Order mirrors the split preference in
-/// [`reorder_tcp_candidates_for_failure`]. Promotion is keyed by id (these span
-/// the `tlsrec_split`, `hostfake`, and `split` families), matching the
-/// id-list approach that function already uses.
-const TLS_VERSION_SPLIT_PREFERRED_IDS: &[&str] = &["tlsrec_split_host", "tlsrec_hostfake_split", "split_host"];
-
-/// Float the SNI-splitting families to the front (in the preferred order),
-/// preserving the relative order of every other spec. Stable and seed-free.
-fn promote_tls_version_split_families(specs: Vec<StrategyCandidateSpec>) -> Vec<StrategyCandidateSpec> {
-    let mut promoted = Vec::with_capacity(specs.len());
-    for id in TLS_VERSION_SPLIT_PREFERRED_IDS {
-        if let Some(spec) = specs.iter().find(|spec| spec.id == *id) {
-            promoted.push(spec.clone());
-        }
-    }
-    for spec in specs {
-        if !promoted.iter().any(|existing| existing.id == spec.id) {
-            promoted.push(spec);
-        }
-    }
-    promoted
 }
 
 pub(in crate::engine::runners::strategy) fn ordered_follow_up_tcp_candidates(
