+++ native/rust/crates/ripdpi-monitor-engine/src/engine/runners/strategy/support/candidate_ordering/tls_ech.rs
@@ -1,0 +1,19 @@
+use crate::types::ProbeResult;
+
+fn probe_detail_value<'a>(result: &'a ProbeResult, key: &str) -> Option<&'a str> {
+    result.details.iter().find(|detail| detail.key == key).map(|detail| detail.value.as_str())
+}
+
+#[cfg(test)]
+pub(in crate::engine::runners::strategy) fn baseline_has_tls_ech_only(results: &[ProbeResult]) -> bool {
+    results.iter().any(|result| result.probe_type == "strategy_https" && result.outcome == "tls_ech_only")
+}
+
+pub(in crate::engine::runners::strategy) fn baseline_supports_ech_candidates(results: &[ProbeResult]) -> bool {
+    results.iter().any(|result| {
+        result.probe_type == "strategy_https"
+            && (result.outcome == "tls_ech_only"
+                || probe_detail_value(result, "tlsEchResolutionDetail") == Some("ech_config_available")
+                || probe_detail_value(result, "cdnProvider").is_some_and(|value| !value.trim().is_empty()))
+    })
+}
