+++ native/rust/crates/ripdpi-tunnel-core/src/io_loop/tcp_accept/admission.rs
@@ -6,22 +6,20 @@ use std::time::Instant as StdInstant;
 use smoltcp::iface::{SocketHandle, SocketSet};
 use smoltcp::socket::Socket;
 use tokio_util::sync::CancellationToken;
-use tracing::info;
 
 use crate::dns_cache::DnsCache;
 use crate::io_loop::packet::{TcpFlowKey, endpoint_to_socketaddr};
 use crate::session::Auth;
-use crate::{ActiveSessions, SessionEntry, Stats};
+use crate::{ActiveSessions, Stats};
 
-use super::duplex::create_session_duplex;
-use super::eviction::remove_evicted_session_socket;
-use super::listener::remove_pending_listen;
-use super::target::{pin_synthetic_ip, pinned_synthetic_ip, tcp_session_target_addr};
+use super::target::{pinned_synthetic_ip, tcp_session_target_addr};
 use super::unresolved::{abort_unresolved_sessions, abort_unresolved_tcp_socket};
 
 mod pending;
+mod session;
 
 use pending::PendingTcpSession;
+use session::admit_session;
 
 /// IANA IP protocol number for TCP, for flow-attribution `note_flow`.
 const PROTO_TCP: u8 = 6;
@@ -80,35 +78,4 @@ fn collect_admissible_sessions(
     }
 
     (new_sessions, unresolvable)
-}
-
-#[allow(clippy::too_many_arguments)]
-fn admit_session(
-    socket_set: &mut SocketSet<'static>,
-    sessions: &mut ActiveSessions,
-    pending_listens: &mut HashMap<TcpFlowKey, (SocketHandle, StdInstant)>,
-    proxy_sockaddr: SocketAddr,
-    auth: &Auth,
-    cancel: &CancellationToken,
-    stats: &Arc<Stats>,
-    dns_cache: &mut Option<DnsCache>,
-    pending: PendingTcpSession,
-) {
-    remove_pending_listen(pending_listens, pending.handle);
-    pin_synthetic_ip(dns_cache, pending.synthetic_ip);
-
-    let session = create_session_duplex(proxy_sockaddr, auth, pending.target_addr, cancel, stats);
-    let entry = SessionEntry {
-        smoltcp_side: session.smoltcp_side,
-        cancel: session.cancel,
-        handle: session.handle,
-        pending_to_session: Vec::new(),
-        pending_to_smoltcp: Vec::new(),
-        upstream_closed: false,
-        pinned_synthetic_ip: pending.synthetic_ip,
-        target_addr: pending.target_addr,
-    };
-    let evicted_handle = sessions.insert(pending.handle, entry);
-    remove_evicted_session_socket(socket_set, evicted_handle);
-    info!("TCP session spawned: remote={}", pending.target_addr);
 }
