+++ native/rust/crates/ripdpi-tunnel-core/src/io_loop/udp_assoc.rs
@@ -1,3 +1,4 @@
+mod association_removal;
 mod association_state;
 mod event_handling;
 mod eviction;
